test: module import error finding working
